requirements.txt
pytesseract==0.3.0
scikit-image==0.15.0
numpy==1.17.2
matplotlib==3.1.1
Pillow==6.2.1
python-dotenv==0.10.3
